rename task classes
